00_demo/pyproject.toml
[project]
name = "00-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.29.5",
    "openai-agents>=0.0.19",
    "pandas>=2.3.3",
    "pandas-stubs>=2.3.2.250926",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
]
